Creating task by User and creating relations with task and user with authentication and bearer tokens.

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from '../auth/user.entity';
 
 /** Use this services to be a model and functions and you can use as an argument any name and them import whatever you have exported */
 
@@ -18,8 +19,11 @@ export class TasksService {
   ) {
   }
 
-  getTasks(filterDto: GetTasksFilterDto) {
-    return this.taskRepository.getTasks(filterDto);
+  getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+  ) {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
   // getAllTasks(): Task[] {
@@ -40,8 +44,8 @@ export class TasksService {
   //   }
   //   return tasks;
   // }
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) {
       throw new NotFoundException(`No task found with the ID of ${id}`);
     }
@@ -56,12 +60,15 @@ export class TasksService {
   //   return found;
   // }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`No task found with the ID of ${id}`);
     }
@@ -79,8 +86,8 @@ export class TasksService {
   //   this.tasks.push(task);
   //   return task;
   // }
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
